requirements.txt
aniso8601==9.0.1
click==8.0.0
dnspython==2.1.0
Flask==2.0.0
Flask-Cors==3.0.10
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
pymongo==3.11.4
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
Werkzeug==2.0.0
